Fix GitHub username regex
The regular expression before only looked for usernames that contained
characters in the \w metacharacter class, which is alphanumerical as
well as underscores. However, this does not match GitHub's allowed
formats for usernames, which is alphanumerical plus hyphens, with the
limitation that it cannot start or end with hyphens, hyphens cannot be
doubled, and a maximum length of 39 characters. Rewrite the regular
expression to these standards to allow for all GitHub usernames to be
linked properly and add a new unit test to ensure it works. Regex
adapted from https://www.npmjs.com/package/github-username-regex.